extract domain objects to common and share common with other projects
