+++ api/src/main/java/dev/local/auth/AuthServiceImpl.java
@@ -2,7 +2,7 @@ package dev.local.auth;
 
 import dev.local.secruity.JwtTokenUtil;
 import dev.local.secruity.JwtUser;
-import dev.local.user.User;
+import dev.local.domain.User;
 import dev.local.user.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
