+++ api/src/main/java/dev/local/role/RoleController.java
@@ -1,5 +1,6 @@
 package dev.local.role;
 
+import dev.local.domain.Role;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Controller;
