+++ api/src/main/java/dev/local/role/RoleRepository.java
@@ -1,5 +1,6 @@
 package dev.local.role;
 
+import dev.local.domain.Role;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
