+++ api/src/main/java/dev/local/tasklist/TaskListController.java
@@ -1,5 +1,6 @@
 package dev.local.tasklist;
 
+import dev.local.domain.TaskList;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
