+++ api/src/main/java/dev/local/user/UserRepository.java
@@ -1,5 +1,6 @@
 package dev.local.user;
 
+import dev.local.domain.User;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
