+++ common/src/main/java/dev/local/domain/TaskList.java
@@ -1,4 +1,4 @@
-package dev.local.tasklist;
+package dev.local.domain;
 
 import lombok.Data;
 import org.springframework.data.annotation.Id;
